Replace the classes for 3.0.3 with 3.1.0
