+++ CHANGELOG.md
@@ -5,9 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.2] - 2025-04-27
+## [1.2.0] - 2025-04-27
 
-- Add OpenAPI Specification v3.1 objects, by @tyzhnenko.
+- Update objects to the OpenAPI Specification v3.1, by @tyzhnenko.
+- Add docstrings to all `v3` classes.
+- Upgrade the `v3.OpenAPI` class from OpenAPI version `3.0.3` to `3.1.0`.
 
 ## [1.1.1] - 2025-03-20
