+++ openapidocs/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.1.1"
+__version__ = "1.2.0"
 VERSION = __version__
